BUG: ProjectionImageFilter did not copy direction cosine information to output
Thi bug is similar to the bug 3475. The direction cosine of the input image
needs to be copied in the output image. It is now copied when the output image
dimension is the same as the input image dimension. If the output image
dimension is smaller (input image dimension minus one), the direction cosine
of the output image is set to the identity matrix.

Change-Id: Ib87e0241a4ea339226ffa117bdbf9b762f525c1f