BUG: LevelSetNeighborhoodExtractor uses spacing
ITK-3482

From the bug report:

"There is a subtle bug in the output of this filter. If you look very closely
at the output level sets, especially if pixel size is not too close to 1,
there is a small stepwise discontinuity right at signed distance = 0 (where
presumably it may be most critical to get accurate values with gradient
magnitude of 1).

This is due to an incorrect formula in itkLevelSetNeighborhoodExtractor.hxx at
line 283. The distance is set to centerValue/(centerValue-neighValue). Clearly
by unit analysis, you can see that this is a unitless value when it should be
in linear spatial units. The correct formula should be distance = centerValue
/ (centerValue - neighValue) * m_InputLevelSet->GetSpacing()[j]."

Change-Id: Id69e9d84c284318167b12e002737eed5ed507a9f