BUG: Template type based input function parameter needs to be deterministic
When instantiating a new object with ITK Python, the template types can be omitted and
will be based on the type of the input parameter.
If the template parameter list contains multiple output types that correspond to the
same input parameter, the template parameters chosen will be the first set of parameters
in the list of keys() that the filter returns.
However, the list of keys() returns by a dictionary has an arbitrary order [1]. This means
that running multiple times the same pipeline could lead to using different sets of
template parameters.
To force a deterministic behavior, we use OrderedDict() instead of Dict().

[1] https://docs.python.org/3/tutorial/datastructures.html

Suggested-by: Bradley Lowekamp <blowekamp@mail.nih.gov>

Change-Id: Ie8cd8ebd6e3fe9d1efe6f25d3a6c47b4b3fcfdc3