ENH: Add OrientedBoudingBox attributes to ShapeLabelMap
This computation is derived from the following contribution:
 "A Label Geometry Image Filter for Multiple Object Measurement"
  by Padfield D., Miller J https://hdl.handle.net/1926/1493

The original algorithm has been made more efficient in the LabelMap
infrastructure with reduced computation only at the start and end of
each RLE line. Additionally, the size of the oriented bounding box has
been extended to include the full extent of each pixel and not just
the center.

Change-Id: I2a4f176b03f2c92780d309f2459f0106fe1d7b48