BUG: include HDF5 config in ITK config
A system HDF5 system may be found in "Module" mode or CMake native
"Config" mode. This patch adds support for HDF5 installed in config
mode where the libraries are exported as targets from a CMake
configuration file, by adding find_package to the installed ITK code.

The variable HDF5_DIR is used to determine which mode the HDF5 package
was found in. When module mode is use this variable is a false value
or "HDF5_DIR-NOTFOUND".

Change-Id: I008935ca918fdcaa8ae24f3e018439ef9face132