ENH: Adds itk_module_examples() macro
This macro simplifies the creation of independent examples in remote modules.
Example folders should contain a project that can be copied outside of their
original project and compiled as is. This means that they will call
find_package(ITK). To keep the example projects as simple as possible, the
'find_package(ITK)' call is run all the time, even if the remote module is
build inside the ITK repository. In this case, the required variables and
configuration files are created as part of the ITK configuration process
and used internally. This is a transparent process from the user's point
of view.

Change-Id: I1b7379d4dfcc183dca31fbe8ab1eeaf38221d727