ENH: Remove 'weak symbol' warning on MacOS
'weak symbol' warnings appear when compiling ITK on MacOS with
BUILD_SHARED_LIBS:BOOL=OFF and ITK_WRAP_PYTHON:BOOL=ON:

ld: warning: direct access in _wrap_itkMersenneTwisterRandomVariateGenerator\
_cast(_object*, _object*) to global weak symbol typeinfo for\
itk::Statistics::MersenneTwisterRandomVariateGenerator means the weak symbol\
cannot be overridden at runtime. This was likely caused by different\
translation units being compiled with different visibility settings.

This is due to export attributes being different some classes when they
are declared and forward-declared. A new export macro called ITK_FORWARD_EXPORT
was created to be used when a class is forward-declared to match the attributes
in the actual declaration of the class.

This macro is not empty only on Mac as it is the only plateform on which
the warnings arise, and adding attributes on the forward declaration
causes other warnings with certain versions of GCC (see commit
79095080f3e69f7014103045e039621fad5a2d83).

In the meantime, GenerateExportHeader.cmake has been simplified and the option
TEMPLATE_EXPORT_MACRO_NAME that had been added has been removed. This option
was used to define template export macros per ITK module. This mechanism has
been replaced by a global ITK_TEMPLATE_EXPORT macro and therefore this
customization of GenerateExportHeader.cmake is not necessary anymore.

Symbol visibility issues have also been corrected by taking advantage of
CMake policy CMP0063 to honor visibility properties for all target types
and not only shared libraries. This is required on MacOS to match
symbol visibility when both static libraries and shared libraries are
compiled and share some symbols (e.g. ITK build statically but with
ITK_WRAP_PYTHON set to ON which builds shared objects).

Missing ITKOptimizers_EXPORT was added on classes declaration.
ITKOptimizers_EXPORT was specified on classes forward declaration, but
not on the classes declaration themselves.

Unnecessary forward declarations that can be replaced by including a header file
were removed. This limits the possible problems of setting different attributes
to the class when it is foward-declared and when it is declared. Having
different attributes creates problems on MacOS.

Finally, some non-templated classes from which templated classes derive need
to always be exported on MacOS and exported as normal non-templated classes
on other systems. Their visibility is set with the macro ITK_FORCE_EXPORT
These classes need to have the same visibility attributes as their derived
classes because these types are used in operations such as dynamic_cast<>().

Change-Id: I5e4b45e4afac988ec68669b28237fa4819214e14