COMP: Use ITK_TEMPLATE_EXPORT in explicit export forward declarations
This is a follow-up to a7d4b0791ea8e56e3d5b2e5b4ab79d93edd91ec1

Use ITK_TEMPLATE_EXPORT in the explicit instantiations. This addresses
'definition of dllimport function not allowed' build errors with shared
build on Windows.

Also change ITK_FORCE_EXPORT to be ITK_FORCE_EXPORT_MACRO where the name
of the module is passed as an argument. This is required for proper dll
export / import macro definitions on Windows.

Change-Id: Ia0386d1301af6aedf6a43940bba5f09109826d82