BUG: Remove weak symbol warnings on MacOS in VNL template classes
VNL template classes are compiled with visibility attributes defined in
VNL_TEMPLATE_EXPORT macro. This macro is not empty only when the libraries
are compiled as shared objects.
This might create weak symbol warnings on MacOS. To remove these warnings,
We manually define the value of this macro if ITK is compiled
statically and with a compiler that sets ITK_TEMPLATE_VISIBILITY_DEFAULT
to ON.

Change-Id: Ida02859798ecf7f2dd7269e2ca9f017667b4dae7