ENH: Add cleanup option to content synchronization script
If new --cleanup option is selected, script verifies that content of MD5 and
SHA512 directory match. This is useful to remove extra files that do not
belong in these folders anymore. However, this is optional as the
external data store might be shared accross projects and files from
other projects may not have a corresponding file for each hash.

For the purpose of the GitHub resource, do not use this new --cleanup flag as
it is important to keep the older files: some are from older revisions of ITK,
and people continue to use the older versions of ITK and request the testing
data.

Change-Id: I673e9404f4f6b9ae1822f71571d05f9f60206ec7