ENH: Add SliceImageFilter
This filter was originally implemented in SimpleITK, and is being
migrated into ITK proper.

This is a new image grid filter which implements functionality similar
to advance array slice indexing found in many scripting
languages. This filter allows for the specification of start, stop,
and stride ( which many be negative ) indexes.

The testing of the filter is implemented to the newly added GoogleTest
library.

Change-Id: I7a8235e25a9008200cfa710d54f66089b9875a9e