ENH: making available GoogleTest for testing in ITK
Added GoogleTest as a third party library. This enables a module to
write a test with the GoogleTest framework as opposed to the classic
ITKTestDriver.

ITK's GoogleTest provides two libraries "ITK::GTest" and "ITK::GTestMain",
which will either be a target or a imported target, which provide the
needed properties for include paths and compilation flags. Therefor,
we do not follow the ITK module convention of defined LIBRARIES or
INCLUDE variables for the module.

Added new CMake macro to create a test driver with gtests:
       CreateGoogleTestDriver

Importing a recent version of FindGTest.cmake and GoogleTest.cmake
from upstream CMake.

Establishing the convention that the test_case_name, the first
parameter to the TEST macro, should be the same as the file name.

Change-Id: I409cf28194a9da9fc5403ca3ecfa15c02c414ad5