ENH: Check for input/output objects with an assertion.
Rather than simply returning when the input/output objects are not set,
check they exist with an assert statement.

Since the processing framework should handle this situation, and throw
the appropriate exception when the input/output objects are not set, the
assertion is preferred over an exception.

Use the itkAssertInDebugAndIgnoreInRelease macro for such purpose.

Change-Id: I5c8181ad9cbb5811cf8f1f229927a0ad89823c4f