BUG: Fixed cylinder bbox, fixed and enhanced IsInside func
The index-to-world transform was not applied to all corner points of the cylinder's bounding box, which caused possible truncation of CylinderSpatialObjects after rotation.
Furthermore, the call to ComputeLocalBoundingBox is now removed from the IsInside function, since calculating the bbox is actually more computational intensive than evaluating the rest of the code.
Added a new test to itkCylinderSpatialObjectTest.cxx for checking the bounding box after rotation.

Change-Id: I05d570d980bcbfb9419b493690571f4575c81478