ENH: Move rank filters out of review
Move the RankImageFilter and the MaskedRankedImageFilter, and the
related histogram classes out of review and into the
MathematicalMorphology module.

While not traditional morphology filters, these filters are
implemented with a structuring element or kernel. Since grayscale
dilate and erode corresponds to the minimum and maximum element,
over a kernel, the rank can be view as a generalization of the
traditional morphology operations.

Change-Id: I7a17c90817bfe038ff49a4f96665ab975cc04f4b