BUG: ITK-3553 fixes ImageOrientation issue in DCMTK reader
A previous reader [1] contained logic to look in two places for
image orientation.

Three years ago a new version [2] was introduced with a logic error.

Before, rval was 0 (success) when reading the regular
ImageOrientationPatient tag and that was returned.
In the new version rval is overwritten by the results of looking for
the per-frame or shared functional groups.

This was introduced when the code was generalized for enhance
multiframes [2].

[1] https://github.com/Kitware/ITK/blob/e081d837a4ae8c7c090456081fdd90610bbe7742/Modules/IO/DCMTK/src/itkDCMTKFileReader.cxx#L1078-L1119
[2] https://github.com/Kitware/ITK/blob/49de8835474d1e49c96cb9565b3c72c8f118b691/Modules/IO/DCMTK/src/itkDCMTKFileReader.cxx#L1096-L1146
[3] https://github.com/Kitware/ITK/commit/98cd612be6f60ec632dc97ac05a64025eb805dba

Change-Id: Iea9e6893ab6e6e3a07cf923d7c04feccae92255d