ENH: Add test from Gaussian distributed samples
Added new test, which uses Google Test, that runs the first order
feature filter on a generated Gaussian sample. This enables validation
of the results based on the expected values from a Gaussian sample. To
compute the tolerance, numpy was used to generate 10,000 Gaussian
distributions, where the distribution of the statistics where
examined. The expected entropy values is computed from expected 101^2
unique samples. A fix seed is used to ensure a repeatable distribution
is generated.