COMP: Fix using system GoogleTest with a compiled system library
Use standard GTest::GTest alias as it is an error to alias an alias.

Use NEW CMP0028 to clarify that ITK's follow the double colon naming
convention for alias.

Do GTest find package int itk-module-init.cmake, to make GTest
variables and alias available to ITK modules which use the module.

Change-Id: I403c72e3bd6bfdec9d1b396c85814d0b51445ac5