BUG: Fix computation of central moments for shape label map
The prior computation, documented in the LabelMap paper (
https://hdl.handle.net/1926/584 ), does not properly
consider the direction cosine matrix. The results were wrong for many
cases of images with directions. The documented formulation assumes
that the points along the run, maintain their position in the non-zero
component.

Details of the re-derivation are found in the code.

Change-Id: Ic3945eb17d073ab55c39026f7bd7ee8e8c612148