ENH: adding explicit testing of computed label shape attributes
Adds testing for bug fixes with the computation of the principal axis,
and the oriented bounding box with direction cosine orientation.

Additionally, this adds comprehensive testing for computation of the
shape label object attributes with the GTest framework. It introduces
a new style of testing where a relatively massive number of values
are easily test with a GTest style macro for array-like items and
mini functions for easy construction of the array-like items.

Move 2D oriented bounding box tests from
itkLabelImageToShapeLabelMapFilterTest2 to the GTest fixture
framework.

Change-Id: Iee8bd44761bd0b2080c5537a6c9e16985d24af79