BUG: Add missing exports for GTest::* targets
When compiling an external ITK module against an ITK build directory
the GTest::* targets were not available. The following CMake
configuration error was generated:

Target "TextureFeaturesGTestDriver" links to target "GTest::GTest" but
the target was not found. Perhaps a find_package() call is missing for
an IMPORTED target, or an ALIAS target is missing?
Call Stack (most recent call first):
test/CMakeLists.txt:259 (CreateGoogleTestDriver)

We are able to maintain the GTest::* namespace for the exported
targets by explicitly setting the export name when building Google
test from ITK source or system supplied source to match that of
CMake's FindGTest.cmake.

Change-Id: I8128f5041466cb3e308113664a31abc29c3f52c6