BUG: Document and assert improper usage of transform methods.
Add documentation of prerequisites for two Transform methods, which
expect that the input and output arguments are different data. If the
same data was passed then the input would be modified during
computation producing incorrect results. An assert is added to aid in
detecting this case during debugging.

The other transform methods have different types for the input and
output so this bug cannot occur with them.

Change-Id: Ia17149ed6cf229ca4c002a5aa9c187d745373c3a