BUG: Fix computation of oriented bounding box
For non-trivial direction cosines, the OBB was not being computed
correctly. Specifically, the step which offset from the center of the
pixels to include the corners.

Now, every vector from the center of the pixel to the corners is
explicitly considered in the principal axis basis to determine the
extrema of the OBB.

Additionally, variables were renamed to clarify the use of the
principal axis as a basis.

Change-Id: I39baa0bb9353b2b468bf63cbae1fe8fecb3b7788