BUG: Propagate make program to KWStyle ExternalProject
Building ITK, configured with the Ninja generator, where the ninja
build program was not in the path resulted in errors when configuring
the KWSytle external project:
  CMake Error: CMake was unable to find a build program corresponding to
  "Ninja".  CMAKE_MAKE_PROGRAM is not set.  You probably need to select
  a different build tool.

The variable "gen" is not defined in the BuildKWStyle.cmake scope, so
the CMAKE_GENERATOR option is removed. Implicitly ExternalProject
propagates the same generator.

Change-Id: Ie4d05d227633f41300b709853d39533055b07857