ENH: Compile ITK with MKL (FFTW) installed on system
To use MKL FFTW implementation:
* Set `ITK_USE_FFTWD`, `ITK_USE_FFTWF`, and `ITK_USE_SYSTEM_FFTW`
to ON and set `ITK_USE_MKL` to ON
* Or set `ITK_USE_MKL` to ON in the CMake command line
* On systems that support TBB threads, `ITK_USE_TBB_WITH_MKL` will
be ON by default. If turned OFF, it uses MKL sequential library.
* Configures `FFTWD_LIB` and `FFTWF_LIB`, but `FFTWF_LIB` is set
to an empty variable because it is not necessary as `FFTWD_LIB`
is enough.
* CMake fatal errors have be added when using system FFTW without
MKL and that `fftw3` or `fftw3f` are not found.
* TBB threads cannot be used with Clang (see [1])
* MKL is linked statically to avoid adding dependencies when linking ITK
to other projects.

[1] https://software.intel.com/en-us/articles/intel-mkl-link-line-advisor

Change-Id: I9c8a4bd57715c28afd9552e2404d898daae33b63