BUG: Fix computation for Jacobian local vector
If the lower triangle of a displacement field was non-zero, then
incorrect jacobians may have been computed.

The ImageBase::TransformLocalVectorToPhysicalVector method must be
called with different input and output vectors or else the input
vector will be clobbered as the output is being generated. This issue
was detected by an recent debug assert added to the method.

Change-Id: I00ffcad6d619d9996293705f63ac47455125749c