ENH: Update to use CircleCI 2.0
Migrate the CircleCI configuration file to the new CircleCI 2.0.

The new platform allows specification of a "large" "resource_class"
computer with 4 cores and 8 GB of memory, which allows compilation and
testing of ITK in ~45min. This removes the need to use distcc over 3
containers. Also additional work for cache of data is no longer
done, it needs to be determined if this would still be beneficial.

Change-Id: Ic3880bb5ba0da9572299a4fa66688e5fd36ffab1