BUG: Rename and replace min/max setter with separate setter methods
For CoocurrenceTextureFeaturesImageFilter:
  m_Min->m_HistogramMaximum respectfully in the
and similarly for the maximum.

For RunLengthTextureFeaturesImageFilter:
  m_Min -> m_HistogramValueMinimum
  m_MinDistance -> m_HistogramDistanceMinimum
and similarly for the maximum.

Add the properly named Set and Get accessors methods according to ITK
convention.

ITK standards are for each parameter or ivar to have a separate set
method. The exists methods which set a min/max pair now call the
standard macro based set methods.