BUG: Use same origin for internal image as input image
The SliceBySliceImageFilter was computing the origin per slice with
the starting index of the image, and not the zero index. This fix
computes the origin based on the zero index at the proper slice. By
computing the origin on a per slice basic, it may changed for each
slice for non-identity direction cosine matrices.

Change-Id: I214f97e6298d78c1857c09655e1fc25333fa8fe9