DOC: Enable offline Doxygen support
Improve Doxygen page load speed and allow offline access to previously visited
Doxygen pages.

This is enabled for browsers that support service workers.

All common page CSS, JavaScript, and PHP resources are pre-cached. HTML,
PNG's, and GIF's are added to the cache as they are visited. The cache is
rotating, and supports 100 entries, which is approximately 30 pages. After
four weeks, it refreshes.

Instructions are added in the comments on how to update the service worker
pre-cache when Doxygen is updated.

Change-Id: I2a1692122196f7ce8eb38090c3aec1d9211860ab