ENH: Remove input name when renamed and add tests
Setting names for index inputs left around old "Primary", "_1",
etc. named inputs after the index was given a name. With this patch
the old name in removed from the list of named inputs.

Additional tests are added for the new AddOptionalInputName
method. These explicitly check the number of named inputs defined and
the behavior of keeping the value of inputs.

Change-Id: I224f796cc1409758e18861d5002ba3c23a864ba6