BUG: itkSimilarity2DTransform was not reporting singular matrices
When inverting a matrix, one has to verify that the determinant is not 0. If
it is 0, the matrix cannot be inverted. In the case of itkSimilarity2DTransform
and itkCenteredSimilarity2DTransform, if the matrix was computed on a singular
matrix, the invert was still computed and GetInverse() function was not
returning "false" to report a problem during computation. This could lead to
invert matrices containing "inf" values without the user to be made aware that
the invert transform was invalid.
Change-Id: I6cb8154683c81e1f876c0e40876072c9a779225f