BUG: Fix links to NOTICE and LICENSE files in README.md.
The links to the NOTICE and LICENSE files were not being rendered as
expected, probably due to the CR/LF that was introduced in a9b44f3 to
avoid the 80 column line width. Markdown splits the link and interprets
that a blank space exists between the hyperlink text and its URL, and
hence it is not rendered correctly.

Change-Id: If327566ebaace5a1588db7e3f277aa843fbf53ea