BUG: sha512 file must exist before complaining that it hasn't been updated
When updating tests files, MD5 content links are created by CMake ExternalData
module. If the MD5 is updated, `pre-commit` verifies that the SHA512 was also
updated. However, if the SHA512 file does not exist, `pre-commit` should not
complain that it hasn't been updated. This case could happen if a user adds
a new data file for which only the MD5 file has been computed, and then
modifies it (or amends the first commit in which the file has been added).

Change-Id: I7aa8abd4f9ae0464f64fbcad413ce9b2e2aeb55a