BUG: Fix marking of required inputs for GPU PDE registration
Recent update to ITK's PDE regisitration class, corrected the
specification of request inputs to explicitly required the FixedImage,
and the MovingImage, instead of incorrectly specify any 2 inputs are
required. The side-effect is when the GPU hierarchy, which appears to
be mostly a copy of the non-GPU PDE hierarchy, resets the number of
required inputs 2 two, the wrong inputs are set to being required.

This patch removed the GPU setting of the number of inputs, and lets
the parent class correctly set the required inputs.