DOC: Document that workaround MSVC 2017 optimization bug has been corrected
In commit [1], a workaround had been introduced to avoid
itkStatisticsAlgorithmTest failing due to a bug in the compiler. This bug has
been corrected in _MSC_FULL_VER == 191125508. A comment is added next to the
workaround in the source code to document the fact that the workaround is not
required for the new MSVC versions.

[1] https://github.com/InsightSoftwareConsortium/ITK/commit/877293170f670e42c5737b86332f9263e8b0b7a1

Change-Id: I426e5e0bf4b9617a90b5074e7b75db0b3d09a8c0