ENH: Extract reusable code into a copy image information functor
Following the design pattern of the ImageRegionCopier, the
ImageInformationCopier is created. This functor operates on ImageBase
objects. For images of the same dimension, it defaults to using the
ImageBase::CopyImageInformation method. For expanding dimensional of
ImageBase objects, the functor fills the extra meta-data with
"identity" information. There is no implementation for the case when
the dimension is being reduced. See the ExtractImageFilter for the
issues related to copying sub-section the direction cosine matrix.

Change-Id: Ic5bd9d6860fa1c5c71f7c6904ea796b012a8ecbf