BUG: Typo in CMake variable name
CMake variable used to declare the byproduct of a project (required by Ninja)
was not always named the same, which means that the variable actually used did
not contain all the values it should have. ITK could not build DCMTK with
Ninja. However, by-products are different for ICU and DCMTK.

Change-Id: I377542518f6c655bb8ce9efa6639898979f9195d