BUG: Support color images (RGB/RGBA) with DCMTK
To enable support for color images with DMCTK, this commit fixes two issues:

(1) a static initialization has to be done by including the header file
"dcmtk/dcmimage/diregist.h". This include was commented out in the ITK code.
The commit reintroduces it.

(2) The color image data are by default organize by plane
(i.e. R1R2R3...G1G2G3...B1B2B3) which is incompatible with the data organization
 that ITK expected, which is by pixel (i.e. R1G1B1R2G2B2R3G3B3...).
To address this, the data has to be copied and reorganized to be usable in ITK.
This reorganization is done by the new function ReorderRGBValues(...) which is
templated over the pixel type.

Details about ReorderRGBValues function:
---------------------------------------

The implementation of this function is based on the function Convert(...) found
in DCMTK (in file dcmimage/include/dcmtk/dcmimage/dicoopxt.h).

This function could be called indirectly in DCMTK by getting the pointer to the
DCMTK data with the function getOutputData() available in the type DicomImage,
instead of using the function getInterData(), but replacing this function call
leads to some tests failing and potential backward incompatibility. Indeed,
this patch only modifies the behavior of itkDCMTKImageIO for color images,
that were previously not working.

Change-Id: I79f17dc2c4fc4f455f48cf9ddcf435863f8282a6