ENH: Improving factory registration
1) Each module can list the factories they want to register using the keyword
`FACTORY_NAMES` when calling `itk_module()`. The factories must have the
following format: <factory>::<format>

2) When the module is loaded, the factories are separated in different lists.
Each factory list contains the list of formats it supports. A list of the
factories that have been configured is created. Finally, a list containing
each format, associated with their factory and the module that declares them
is created (<module>::<factory>::<format>).

3) Automatic factory registrations can be performed with other types of classes
then I/O.

4) At configuration time, a project that uses ITK can select which format
should be registered and in which order.

5) Remote modules can be seamlessly registered the same way other ITK modules
are.

Change-Id: Ieed2698d08f9c7b01fd03b30df69315d5da0e383