ENH: Adding factory components
When calling `find_package()` with a list of components, it is now possible
to not list each factory individually, but to direcly load them all for
each factory type (e.g. ImageIO, TransformIO). To load all components of a
factory, add the component `ITK<factory>` in the list of requested components.

Change-Id: I6cfa15adaa9bdb18d8be8a71e3a6e3d9a1203580