ENH: Add __call__() function in Python itkTemplate objects
itkTemplate objects can be run directly as a function without having to
instantiate an object before. This allows a more functional usage of
ITK Python. The __call__() function runs the filter on the largest possible
region and returns the output of the filter.

e.g.: image = itk.MedianImageFilter(input_image, Radius = 2)

Change-Id: I386a217ac81949621872fd6a3c058a45f831edc5