BUG: Python docstrings were not generated anymore
For each ITK module, the generation of the Python docstrings is done in two
stages:
1) XML files containing the doxygen documentation is generated.
2) From these XML files, *_doc.i files are then created and incorporated by
SWIG in the Python wrapping.

This mechanism was broken:
1) Newer CMake version requirements changed the behavior of how variables are
interpreted in `if` statements (CMP0054).
2) Doxygen output file names has changed from all lower case to snake case
(additional '_' between words).

Some additional minor modifications have been incorporated in this patch:
* Remove outdated code to print Python documentation. This is not used anymore
as the documentation is directly incorporated in the Python docstrings.
* Improved itkTemplate documentation.
* `ITK_WRAP_DOC` was not saved in `ITKConfig.cmake`.

Note: To build the documentation, `ITK_WRAP_DOC` has to be set to ON when
configuring ITK.

TODO: This patch does not address the following warnings:
"ninja: warning: multiple rules generate"
This will have to be addressed in a follow-up patch.

Change-Id: Ic3cdea797e1a05005228431f6f259acb0a33f5cc