BUG: HoughTransform CirclesList should be empty when NumberOfCircles is zero
When a user explicitly does filter->SetNumberOfCircles(0),
itk::HoughTransform2DCirclesImageFilter::GetCircles() should return an empty
list of circles.

Note that most use cases of the function probably won't be affected by this
fix, as m_NumberOfCircles is initialized to one in the default-constructor of
the filter, and a user most probably just calls SetNumberOfCircles(n) to
increase the number.

Change-Id: I1796757815de73a4afec70c3f3da39fcba4f154f