BUG: Disable checking of OBB origin
The solution for the OBB is not unique due to ambiguity in the sign of
the eigen vectors for the axes. This particular case does not produce
the same results on 32-bit and 64-bit intel system due to floating
point tolerances. We are just disabling test for the problematic
case.

Change-Id: I6b2fa002dc799df2b7b786ebd478bde453c0a53b