BUG: Add previously removed function __getattr__() in itkTemplate.py
The function __getattr__() had been removed in commit [1] because it was
documented as being implement for docstring support and was not necessary
anymore for this purpose. However, it is also important to allow lazy loading
to work correctly. Part of this function has now been reimplemented to
correct this issue.

Without this function, types that are defined in a module but do not belong
in that module are not loaded with the operator '.', but only with the operator
'['. This is the case for example with images of type NormalBandNode that are
not defined in the ITKCommon module, but in the ITKLevelSets module.

[1] d7deda60f7145c89c4df0bdb66627e4b8142b9ce

Change-Id: I97d7930516246005771bdc44bf6dee9c63596e6e