BUG: Fix LBFGS2 iteration count and iterface
Change the 1-base iteration count in libLBFGS to ITK's convention of
the first iterations being 0. Match the behavior of other optimizer,
when the maximum number of iterations are reached then the last
current iteration value is the maximum stopping criteria.

Add mapping of base class Set/GetNumberOfIterations methods to
LBFGS2's "MaximumIteration" method to comply with base class
interface.

Change-Id: I37b84cbee4a54d6e4e48ce19ffd68572ac27af7b