ENH: Create virtual interface for the MultiTheader
This allows separate implementation of multithreaders. Perhaps for
conventional threading, thread pool, or some other threading library
or model. The creation of the proper MultiThreaders can occur in the
MultiThreaderBase::New method based on global defaults. Or a filter
could manually specify/set the desired multithreader.

Change the ProcessObject interface to use the MultiThreaderBase
interface.

Change-Id: I9e136f13e4e8c8b9c54246d5084c1b45b3c6d07c