STYLE: HoughTransform calling convenience overload of SetVariance
Simplified gaussianFilter->SetVariance calls in GetCircles() and GetLines()
by using the SetVariance(ValueType) convenience overload.

Removed unnecessary 'typename' from InternalImageType::Pointer, in
GetCircles(), as this type name is not dependent on the template parameters
of HoughTransform2DCirclesImageFilter. Removing the 'typename' here is in
accordance with the corresponding variable in GetLines().

Added 'const' to a few more local variables.

Change-Id: I78eb95e11a1e68f2cf0ca92b2e2d74e34657c5eb