ENH: Allow different types accumulator and radius image Hough Circles filter
Before this commit, the accumulator image and the radius image generated by
itk::HoughTransform2DCirclesImageFilter both had the same pixel type,
specified by template parameter TOutputPixelType. It would be better to allow
these two images to have different pixel types: for the accumulator pixel
type, an unsigned integer type is usually the best choice, whereas for the
radius pixel type, 'double' appears most appropriate.

This commit solves this issue by adding TRadiusPixelType as a template
parameter, defaulting to TOutputPixelType for backward compatibility.

A test is added to check that the choice for TRadiusPixelType does not affect
the estimated centers of the found circles. The estimated radii may get
slightly higher values when changing TRadiusPixelType from an integer type
to a floating point type, because floating point radius estimations are not
being truncated.

Change-Id: I21894dc27bb7d3334ee5ce0de2ad1615713ef3a8