COMP: UseITK: Fix configure error due to different subproject requirements
If within a project depending on ITK two subprojects A and B requires
different IO factories by calling:

  find_package(ITK COMPONENTS <list_of_modules> <list_of_io_factory> REQUIRED)
  include(${ITK_USE_FILE})

where <list_of_io_factory> is different for both projects. The second
project will use the list "Factory_${_factory_name}_formats" from the
first project and a "Module not found" error will be reported.

This commit addresses the problem by not caching the Factory_${_factory_name}_formats
variable.

The error was encountered while updating Slicer to build against an updated
ITK.

As a result of this patch, the correct way to specify which formats will be
statically registered when configuring a project is to specify the IO
components to load when calling
`find_package(ITK COMPONENTS <list_of_modules>)`.

Specific IO modules can be added to the list of modules even when using
lists of IO factories to enforce the requirement of these specific modules:

e.g.: `find_package(ITK COMPONENTS ITKCommon ITKImageIO ITKIONRRD REQUIRED)`

Change-Id: I3ad07d295bf16b7488b53e928e30d3b366d5aaa5