BUG: Respect ITK_FUTURE_LEGACY_REMOVE
The command line needs to respect the flag

cmake -DITK_FUTURE_LEGACY_REMOVE:BOOL=ON .
-- or --
cmake -DITK_FUTURE_LEGACY_REMOVE:BOOL=OFF .

to induce compile time failures when external
packages use deprecated features.

the current implementation always sets this
value to 0 in the CACHE, thereby preventing
the ability to test for use of features that
are slated to be removed in future versions.

Change-Id: I343b58091f953e7d667343f6426ee6788bf57b28