BUG: Access violation itkGaussianDerivativeOperatorTest argv[0]
I encountered a crash on the following scenario:
 - I ran ITKCommon1TestDriver.exe
 - It prompted: "To run a test, enter the test number:"
 - I entered '12', to run itkGaussianDerivativeOperatorTest
 - Output: Access violation reading location 0x0000000000000000

In this scenario, ITKCommon1TestDriver.cxx decrements the number of arguments
from main(ac, av), from 1 to 0.
The test (itkGaussianDerivativeOperatorTest.cxx) treats zero arguments
(argc == 0) as an error, and then tries to print argv[0], which is a null
pointer, in this scenario. So then it crashes.

Instead, itkGaussianDerivativeOperatorTest should just start testing
itk::GaussianDerivativeOperator!

This commit fixes the bug as it now only treats argc > 1 as error.

Change-Id: I4845ae7f3a71313b90fcf2325b67e161139493dc