BUG: Allow compilation of VNL with MSVC<1600
Compiler detection and feature support is performed with files generated with
the CMake macro `WriteCompilerDetectionHeader`. This macro generate a specific
header file for several supported compilers, including MSVC. However, for
MSVC, it only supports feature detection if MSVC >= 1600. If MSVC < 1500,
an error message is thrown instead of defaulting to default unknown compiler
which uses `__cplusplus` variable to configure supported features. In the case
of MSVC < 1600, it will most likely default to a configuration with no
C++11 support.

Change-Id: I45067bbec103f7e24ecb384c85c70b065a450503