BUG: Create IO Meta modules at module loading time
Instead of sorting through the meta-modules such as ITKImageIO and
ITKTransformIO in ITKConfig.cmake, to remove them from the list of
requested module, we create a meta-module that is marked as LOADED
and that lists its dependencies (all the modules invoking the
corresponding factory) in TRANSITIVE_DEPENDS modules.
This should allow developers to call `itk_module_config()` with
meta-modules.

This problem was mentioned in SimpleITK/SimpleITK#348

Change-Id: I88e317562b46c4a0c7b0f0fe361f4d9d645463fd