BUG: Keep the SetSeed# methods in IsolatedConnectedImageFilter
SimpleITK is using these methods. A method to explicitly set one seed,
as oppose to appending to the current list is useful because it
explicitly sets the state of the IVAR without dependency of the state
of the IVAR.

In general ITK does not have consistent interfaces with how "seed" are
managed, and what the expected interface should be. This is future work.

Change-Id: I097dbd2e9264d69eb2a1f4e7aa4be426e24bec86