ENH: Revert to include itkv3::Rigid3DTransform
This class is primarily needed to ease migration.  ITKv4 attempted to remove
this class completely, but unfortunately several work-arounds were used
subsequently added that essentially provided the itkv3 interface.

By default the previous itkv3 interface is provided, but
this is marked to be removed in a future version of ITK.

Change-Id: I32cfb6a2c37a03835485721e90917ed6d76604af