BUG: Provide consistent GetOutput behavior
The interface for GetOutput that mimics
the behavior of ProcessObjects was confounded
with the inferace for returning ivars.  These
two API's are incompatible with respect to
const return values.

The ivars names were changed from m_Object to
make it more clear that the intent is to provide
the GetOutput interface, instead of exposing
the ivar through the Get/Set interface paradigm.

The itk::ImageSource api was used as a reference
  OutputImageType * GetOutput(); 
  const OutputImageType * GetOutput() const;

The GetModifiableObject() interface was exposed
for a long time in ITKv4, so it is retained to
provide backwards compatibilty support as a legacy
Application.

Change-Id: I93df321f7b48d42542ec776d7b92fb8ce39c8bbb