BUG: GaussianDerivativeImageFunction should use image spacing consistently
Moved recomputation of Gaussian kernels from the default-constructor of
GaussianDerivativeImageFunction to SetInputImage, as the computation of these
kernels depends on the spacing of the input image.

Note: This looks like a serious bug. Previously, the image spacing was
sometimes ignored, even though m_UseImageSpacing was always true. This would
have affected the results of the Hough Transform Circles filter, depending on
itk::HoughTransform2DCirclesImageFilter::m_SigmaGradient.

Change-Id: Ic8a2233ef552569481375ebbedb4a4a6658e905e